bugs fix - uploaded to godaddy

+++ Retrieval/deleteFile.php
@@ -1,39 +1,43 @@
 <?php
+    if(isset($_GET['fileid'])) {
+        $fileId = $_GET['fileid'];
 
-
-    $fileId = $_GET['filename'];
-
-        // craate sql connection
-        include('/includes/connection.php');
+        // create sql connection
+        include('includes/connection.php');
 
         // make query to sql files table
-        $query = "select * from Files";
+        $query = "SELECT `fileName` FROM Files WHERE `fileID`='".$fileId."';";
 
         // send query to sql files table
         $result = mysqli_query($connection , $query);
         if( !$result ){
-            echo "DB query failed from file_bars.php";
-            die("DB query failed from file_bars.php");
+            die("DB query failed from delete.php : " . mysqli_connect_error());
         }
-
-        // print all sql data rows
-        while( $row = mysqli_fetch_assoc($result) ){
-            if($fileId==$row["fileID"])
-                $filename=$row["fileName"];
-        }
-
+        $row = mysqli_fetch_assoc($result);
+        $filename =  $row['fileName'];
 
         $chars = explode("../data/",$filename);
         $convert = implode("",$chars);
         $temp = '/data/';
         $final = $temp.$convert;
 
-    $deletedfile = getcwd()."".$final."";
+        $deletedfile = getcwd()."".$final."";
 
-    unlink($deletedfile);
+        unlink($deletedfile);
 
-    // close sql connection
-    mysqli_free_result($result);
-    mysqli_close($connection);
 
+        $query = "DELETE FROM Files where `fileID`='".$fileId."';";
+        $result = mysqli_query($connection , $query);
+        if( !$result ){
+            die("DB query failed from delete.php : " . mysqli_connect_error());
+        }
+
+        // close sql connection
+        mysqli_free_result($result);
+        mysqli_close($connection);
+        header('Location: ../intro.html');
+
+    } else {
+        echo 'Cant fetch filename';
+    }
 ?>
